✨Refactor BaseEntity to Support Multiple Id Types
