+++ src/Domain/Common/BaseAuditableSoftDeleteEntity.cs
@@ -1,0 +1,11 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace CleanArchitecture.Blazor.Domain.Common;
+
+public abstract class BaseAuditableSoftDeleteEntity : BaseAuditableEntity, ISoftDelete
+{
+    public DateTime? Deleted { get; set; }
+    public string? DeletedBy { get; set; }
+
+}
