+++ src/Domain/Entities/Audit/AuditTrail.cs
@@ -7,9 +7,9 @@ using Microsoft.EntityFrameworkCore.ChangeTracking;
 
 namespace CleanArchitecture.Blazor.Domain.Entities.Audit;
 
-public class AuditTrail : IEntity
+public class AuditTrail : IEntity<string>
 {
-    public int Id { get; set; }
+    public string Id { get; set; } = Guid.NewGuid().ToString();
     public string? UserId { get; set; }
     public virtual ApplicationUser? Owner { get; set; }
     public AuditType AuditType { get; set; }
