+++ src/Domain/Entities/Document.cs
@@ -7,7 +7,6 @@ namespace CleanArchitecture.Blazor.Domain.Entities;
 
 public class Document : OwnerPropertyEntity, IMayHaveTenant, IAuditTrial
 {
-    public int Id { get; set; }
     public string? Title { get; set; }
     public string? Description { get; set; }
     public JobStatus Status { get; set; } = default!;
