+++ src/Domain/Entities/Product.cs
@@ -5,7 +5,6 @@ namespace CleanArchitecture.Blazor.Domain.Entities;
 
 public class Product : BaseAuditableEntity
 {
-    public int Id { get; set; }
     public string? Name { get; set; }
     public string? Description { get; set; }
     public string? Brand { get; set; }
