+++ src/Domain/Entities/Tenant.cs
@@ -1,5 +1,5 @@
 namespace CleanArchitecture.Blazor.Domain.Entities;
-public class Tenant : BaseAuditableEntity
+public class Tenant : IEntity<string>
 {
     public string Id { get; set; } = Guid.NewGuid().ToString();
     public string? Name { get; set; }
