+++ src/Domain/Events/DeletedEvent.cs
@@ -3,7 +3,7 @@
 
 
 namespace CleanArchitecture.Blazor.Domain.Events;
-public class DeletedEvent<T> : DomainEvent where T : BaseEntity
+public class DeletedEvent<T> : DomainEvent where T : IEntity
 {
     public DeletedEvent(T entity)
     {
