+++ src/Domain/Events/UpdatedEvent.cs
@@ -1,5 +1,5 @@
 namespace CleanArchitecture.Blazor.Domain.Events;
-public class UpdatedEvent<T> : DomainEvent where T : BaseEntity
+public class UpdatedEvent<T> : DomainEvent where T : IEntity
 {
     public UpdatedEvent(T entity)
     {
