+++ src/Infrastructure/Persistence/Configurations/TenantConfiguration.cs
@@ -9,6 +9,6 @@ public class TenantConfiguration : IEntityTypeConfiguration<Tenant>
 {
     public void Configure(EntityTypeBuilder<Tenant> builder)
     {
-        builder.Ignore(e => e.DomainEvents);
+        //builder.Ignore(e => e.DomainEvents);
     }
 }
